% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash_cor.R
\name{ash_cor}
\alias{ash_cor}
\title{Adaptive shrinkage of a correlation matrix}
\usage{
ash_cor(cormat, nsamples, image = FALSE)
}
\arguments{
\item{cormat}{The estimated sample correlation matrix}

\item{nsamples}{The number of samples over which the correlation matrix is estimated.}

\item{image}{if TRUE, plots an image of the shrunk and non-shrunk correlation matrices}
}
\value{
Returns a shrunk version of the sample correlation matrix (the output is also a correlation matrix)
}
\description{
This function performs adaptive shrinkage of a correlation matrix
}
\keyword{adaptive}
\keyword{correlation}
\keyword{shrinkage,}

