% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash_cov_master.R
\name{ash_cov_master}
\alias{ash_cov_master}
\title{Adaptive shrinkage of correlation matrix from data}
\usage{
ash_cov_master(data)
}
\arguments{
\item{data}{The samples by features data matrix}
}
\value{
The adaptive shrinkage correlation and covariance matrices, the shrink intensity and the Ledoit-Wolf adjusted adaptive shrunk versions of correlation and covraince matrices.
}
\description{
This function performs adaptive shrinkage of sample correlation matrix obtained from data
}
\keyword{adaptive}
\keyword{correlation}
\keyword{shrinkage,}

