% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shafer_strimmer_shrinker.R
\name{shafer_strimmer_shrinker}
\alias{shafer_strimmer_shrinker}
\title{Performs Shafer-Strimmer shrinkage on covariance matrix}
\usage{
shafer_strimmer_shrinker(data)
}
\arguments{
\item{data}{The samples by features data matrix.}
}
\value{
Returns a list with two items - the Shafer-Strimmer shrunk covariance matrix and the shrinkage intensity.
}
\description{
This function performs Shafer Strimmer shrinkage on a covariance matrix
}
\keyword{Shafer-Strimmer}
\keyword{shrinkage}

